requirements.txt
flask
torch
torchvision
numpy
pandas
pillow
opencv-python
